requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
loguru==0.6.0
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
yarl==1.8.2
